Use local CSS rules for CodeView
Instead of having a global appearance for the CodeView widget chosen
at application level, this commit will make each CodeView have its own
CSS rules for setting the font appearance.

This makes use of the deprecated StyleContext class, which will be
removed in GTK 5. However, for the intended use case (zoom) there is
no alternative at the moment. After checking out how other
applications do this, it turns out that GNOME Text Editor is using CSS
for the zoom settings, and they use these deprecated methods. Given
that GNOME Text Editor is a core app of GNOME, I bet that they won't
release GTK 5 without a solution to allow apps to customize styles
like this. So I'll just continue using the deprecated methods and if
other apps start to switch to a new way to set zoom level or inline
CSS rules, I'll just mimic what they do.
