+++ data/ui/code_export_pane.blp
@@ -29,10 +29,6 @@ template $CarteroCodeExportPane: $CarteroBaseExportPane {
         hexpand: true;
 
         $CarteroCodeView view {
-          styles [
-            "use-cartero-font",
-          ]
-
           left-margin: 10;
           right-margin: 10;
           top-margin: 10;
