+++ data/ui/raw_payload_pane.blp
@@ -25,10 +25,6 @@ template $CarteroRawPayloadPane: $CarteroBasePayloadPane {
       vexpand: true;
 
       $CarteroCodeView view {
-        styles [
-          "use-cartero-font",
-        ]
-
         editable: bind template.read-only inverted;
         top-margin: 10;
         bottom-margin: 10;
