+++ data/ui/response_panel.blp
@@ -59,10 +59,6 @@ template $CarteroResponsePanel: Adw.Bin {
                 vexpand: true;
 
                 $CarteroCodeView response_body {
-                  styles [
-                    "use-cartero-font",
-                  ]
-
                   top-margin: 10;
                   bottom-margin: 10;
                   left-margin: 10;
